fix(CIRC-10300): Ensure SNMP check OID settings do not swap places
Change the type of the OID configurations for SNMP checks to use
TypeSet, rather than TypeList. This ensures that the state is always
ordered in the same way, regardless of how the API returns the
configuration keys.

This should resolve the issue of OID's changing positions in the state
when applying SNMP checks.